Titanic Project
This is my first project of Data Science. I took time to understand how to work and follow some commom steps about dealing with data. I enrolled on some courses, but the best way to learn is practicing, because of this I started with the famous project "Titanic Survival Prediction". There are a lot to improve, but this is the first version.